20230116/frontend/requirements.txt
flask
numpy
python-dateutil
